Enemy AI Merge Patch
Fixed some issues with object not merging correctly from the last AI Merge.
This includes some tag names, map missing items, prefab spawn lists and a few others.
More testing needs to be done to confirm that everything from the merge is working correctly but for now it seems to be ok.