Materials and structures synced on server and client
Players can now change materials and structure types using keys 1-3 to change between wood, stone and metal. 4-7 change between wall, stairs, doorway and door.

When a player spawns a structure the spawned structure will be the same type and material chosen and will be visible on both clients and server.