Completely rebuilt base building system
Base building has been built from the ground up over the past week. The previous version was networked but did not allow for a base building mode as well a placement ghost and material and structure choices being persistent between structure placement in the player. This is now all implemented as well as the correct networking. client side visualisation of the changes needs to be tweaked dto show correctly but the system is functioning as intended now.