Weapon Class Slots and Switching
Weapon slots have now been added in, these allow for generic weapons to be placed into them and selected for firing. This will allow for a wide range of weapons to be made and added at any point. The weapon switiching has also been added and networked so that its synced across all clients and server.