Added Trap prefab as well as logic code
Added the prefab and neccessary colliders. Also added code to execute a timer that resets while the enemy is colliding with the trap. This will cycle and do damage while the enemy is still colliding. Code has been implemented to slow the enmies down while they're in the traps radius, although this isn't confirmed to be working yet.