Added a few more audio files into scripts
Local sounds based at the player can be done quite easily.
Links in with the audio mixer.
However, sounds played at different positions or over the network are not so easy.
Do not link with the audio mixer correctly and do not play on the correct connection
Needs more research and testing to get working correctly.