Basic AI Spawning Added
The director will now spawn AI units near the player based on the Day/Night bool as well as the player intensity. This is still lacking many of the more advanced spawning parameters but shows that the system works in its most basic state.
The debug overlay has also been updated to show AI information