Corrected audio replicaiton over network
To get the audio working correctly there is now a middle-man audio controller. This is held as a reference on an object that needs its audio networked. The object calls this controller which then has access to an audio source on a player in which the sound is replicated to.

Local audio can still be simply carried out.