Controller Aiming Deadzone remade
The deadzone settings for the aiming has been remade to not use the built in unity settings. This allows for a vector to be used as a deadzone, rather than each axis seperatly. This reduces the possibility of an axis zeroing out before the other, which should stop the players from snapping to 0,90,180 or 270 degrees when releasing the stick.