Framework change
Weapon audio appears to be based on FPS due to the timing used in the shoot method. Considering separating this out.
Changed the audio framework to account for a canClip boolean, checks whether the audio file should check if one is currently playing or not. Ideal for the weapon shooting.
Fixed the audioSync controller not being assigned on play.