Particle System Update for shooting
The particle system used for the player shooting has been rebuilt and now provides collision testing for the particles fired. This now stop the particles from passing through objects that they should not.

This partially addresses issue #126 but still not not seem to check for dynamic colliders within the scene.