AI base targetting rebuild
Due to the performance issue with the previous buidling detection method, the location of targetable buildings has now been transferred to the AI directior. This means that when the AI need to target a building they send a request to the AI director who then provides them with a target to attack. The director istself will generate a list of targetable objects when it spawns a wave of AI and also manages deletion of the targetable objects.
Using this method has remove all stuttering encountered by the AIs previous pathfinding method, but still requires further testing.