Director base detection update
This push brings a number of updates that aim to handle a number of null references that have been causing numerous performacne issues and crashes caused by the new system and  networking ti with multiple players.
This system appears to be somewhat stable now, but again will require more testing before merging.