AI Director Day/Night integration with day/night script
This commit contains changes to the AIs pathinfincing that should reduce the amount of unneeded pathfinding calculations .

Also contains additions to the day/night script to sync the day and night time with the director.  Meaning that the AI Director now, automatically switches from day to night spawning mode according to the time within the day night script.