Day Night Update
The time link between day/night and the director was causing issue on clients as they could not set a reference to the director as it only exisits on the server. The clients have now been updated to not attempt any interaction with the director through thed ay/night cycle.